Modified Components to pass in AppState

+++ src/Window.zig
@@ -3,37 +3,53 @@ var gl_procs: gl.ProcTable = undefined;
 
 pub var current: ?*Self = null;
 
-window: glfw.Window,
+window: blk: {
+    if (builtin.is_test) {
+        break :blk EmptyStruct;
+    } else {
+        break :blk glfw.Window;
+    }
+},
 ctx: Component,
 graphics: gfx.Graphics,
 font_lib: ft.Library,
 
 /// Call `Window.init()` before using the Window API
 pub fn init() Err!void {
+    if(builtin.is_test) {
+        return;
+    }
     if (!glfw.init(.{})) return Err.FAILED_TO_INITIALIZE_GLFW;
 }
 
 pub fn new(allocator: std.mem.Allocator, title: [*:0]const u8, width: u32, height: u32) !Self {
-    const window = glfw.Window.create(width, height, title, null, null, .{
-        .decorated = false,
-        .context_version_major = 2,
-        .context_version_minor = 0,
-        .samples = 4,
-        .auto_iconify = true,
-    }) orelse return Err.FAILED_TO_CREATE_WINDOW;
-    errdefer window.destroy();
+    const window = blk: {
+        if (builtin.is_test) {
+            break :blk EmptyStruct{};
+        } else {
+            break :blk glfw.Window.create(width, height, title, null, null, .{
+                .decorated = false,
+                .context_version_major = 2,
+                .context_version_minor = 0,
+                .samples = 4,
+                .auto_iconify = true,
+            }) orelse return Err.FAILED_TO_CREATE_WINDOW;
+        }
+    };
+    if (!builtin.is_test) {
+        errdefer window.destroy();
 
-    glfw.makeContextCurrent(window);
-    window.setInputModeStickyKeys(true);
+        glfw.makeContextCurrent(window);
+        window.setInputModeStickyKeys(true);
 
-    if (!gl_procs.init(glfw.getProcAddress)) return error.FAILED_TO_INITIALIZE_OPENGL;
-    gl.makeProcTableCurrent(&gl_procs);
+        if (!gl_procs.init(glfw.getProcAddress)) return error.FAILED_TO_INITIALIZE_OPENGL;
+        gl.makeProcTableCurrent(&gl_procs);
 
-    gl.Disable(gl.DEPTH_TEST);
-    gl.Enable(gl.BLEND);
-    gl.BlendFunc(gl.SRC_ALPHA, gl.ONE_MINUS_SRC_ALPHA);
-    gl.Enable(gl.TEXTURE_RECTANGLE_ARB);
-
+        gl.Disable(gl.DEPTH_TEST);
+        gl.Enable(gl.BLEND);
+        gl.BlendFunc(gl.SRC_ALPHA, gl.ONE_MINUS_SRC_ALPHA);
+        gl.Enable(gl.TEXTURE_RECTANGLE_ARB);
+    }
     const graphics = try gfx.Graphics.new(allocator, width, height);
     errdefer graphics.destroy();
 
@@ -53,63 +69,65 @@ pub fn new(allocator: std.mem.Allocator, title: [*:0]const u8, width: u32, heigh
 }
 
 pub fn resize(self: *Self, width: u32, height: u32) !void {
-    gl.Viewport(0, 0, @intCast(width), @intCast(height));
-    gl.MatrixMode(gl.PROJECTION);
-    gl.LoadIdentity();
-    gl.Ortho(0.0, 1.0, 0.0, 1.0, -1.0, 1.0);
+    if (!builtin.is_test) {
+        gl.Viewport(0, 0, @intCast(width), @intCast(height));
+        gl.MatrixMode(gl.PROJECTION);
+        gl.LoadIdentity();
+        gl.Ortho(0.0, 1.0, 0.0, 1.0, -1.0, 1.0);
 
-    gl.Clear(gl.COLOR_BUFFER_BIT);
+        gl.Clear(gl.COLOR_BUFFER_BIT);
 
-    gl.DeleteTextures(1, @ptrCast(&_internal_texture_id));
-    gl.GenTextures(1, @ptrCast(&_internal_texture_id));
-    gl.BindTexture(gl.TEXTURE_RECTANGLE_ARB, _internal_texture_id);
-    gl.TexImage2D(gl.TEXTURE_RECTANGLE_ARB, 0, gl.RGBA, @intCast(width), @intCast(height), 0, gl.BGRA, gl.UNSIGNED_BYTE, null);
-    gl.TexEnvi(gl.TEXTURE_ENV, gl.TEXTURE_ENV_MODE, gl.DECAL);
-
+        gl.DeleteTextures(1, @ptrCast(&_internal_texture_id));
+        gl.GenTextures(1, @ptrCast(&_internal_texture_id));
+        gl.BindTexture(gl.TEXTURE_RECTANGLE_ARB, _internal_texture_id);
+        gl.TexImage2D(gl.TEXTURE_RECTANGLE_ARB, 0, gl.RGBA, @intCast(width), @intCast(height), 0, gl.BGRA, gl.UNSIGNED_BYTE, null);
+        gl.TexEnvi(gl.TEXTURE_ENV, gl.TEXTURE_ENV_MODE, gl.DECAL);
+    }
     try self.graphics.resize(width, height);
 }
 
 /// Render the Window and update the UI
-pub fn update(self: *Self) !void {
-    gl.MatrixMode(gl.MODELVIEW);
-    gl.LoadIdentity();
-    gl.Clear(gl.COLOR_BUFFER_BIT);
+pub fn update(self: *Self, app: *App) !void {
+    if (!builtin.is_test) {
+        gl.MatrixMode(gl.MODELVIEW);
+        gl.LoadIdentity();
+        gl.Clear(gl.COLOR_BUFFER_BIT);
 
-    gl.PushMatrix();
+        gl.PushMatrix();
 
-    gl.BindTexture(gl.TEXTURE_RECTANGLE_ARB, _internal_texture_id);
-    gl.TexImage2D(gl.TEXTURE_RECTANGLE_ARB, 0, gl.RGBA, @intCast(self.graphics.surface.width), @intCast(self.graphics.surface.height), 0, gl.BGRA, gl.UNSIGNED_BYTE, @ptrCast(&self.graphics.surface.data[0]));
+        gl.BindTexture(gl.TEXTURE_RECTANGLE_ARB, _internal_texture_id);
+        gl.TexImage2D(gl.TEXTURE_RECTANGLE_ARB, 0, gl.RGBA, @intCast(self.graphics.surface.width), @intCast(self.graphics.surface.height), 0, gl.BGRA, gl.UNSIGNED_BYTE, @ptrCast(&self.graphics.surface.data[0]));
 
-    gl.Color3f(0, 1, 0);
-    gl.Begin(gl.QUADS);
+        gl.Color3f(0, 1, 0);
+        gl.Begin(gl.QUADS);
 
-    gl.TexCoord2f(0, @floatFromInt(self.graphics.surface.height));
-    gl.Vertex2f(0, 0);
+        gl.TexCoord2f(0, @floatFromInt(self.graphics.surface.height));
+        gl.Vertex2f(0, 0);
 
-    gl.TexCoord2f(@floatFromInt(self.graphics.surface.width), @floatFromInt(self.graphics.surface.height));
-    gl.Vertex2f(1, 0);
+        gl.TexCoord2f(@floatFromInt(self.graphics.surface.width), @floatFromInt(self.graphics.surface.height));
+        gl.Vertex2f(1, 0);
 
-    gl.TexCoord2f(@floatFromInt(self.graphics.surface.width), 0);
-    gl.Vertex2f(1, 1);
+        gl.TexCoord2f(@floatFromInt(self.graphics.surface.width), 0);
+        gl.Vertex2f(1, 1);
 
-    gl.TexCoord2f(0, 0);
-    gl.Vertex2f(0, 1);
+        gl.TexCoord2f(0, 0);
+        gl.Vertex2f(0, 1);
 
-    gl.End();
+        gl.End();
 
-    gl.PopMatrix();
-    self.window.swapBuffers();
-    glfw.pollEvents();
+        gl.PopMatrix();
+        self.window.swapBuffers();
+        glfw.pollEvents();
 
-    const frame_size = self.window.getFramebufferSize();
-    if (frame_size.width != self.graphics.surface.width or frame_size.height != self.graphics.surface.height) {
-        try self.resize(frame_size.width, frame_size.height);
+        const frame_size = self.window.getFramebufferSize();
+        if (frame_size.width != self.graphics.surface.width or frame_size.height != self.graphics.surface.height) {
+            try self.resize(frame_size.width, frame_size.height);
 
-        self.ctx.invalidate();
+            self.ctx.invalidate();
+        }
     }
-
-    try Context.update(&self.ctx);
-    try Context.syncAndCalculateSize(&self.ctx, &self.graphics);
+    try Context.update(&self.ctx, app);
+    try Context.syncAndCalculateSize(&self.ctx, app, &self.graphics);
 }
 
 pub fn shouldClose(self: Self) bool {
@@ -117,11 +135,17 @@ pub fn shouldClose(self: Self) bool {
 }
 
 pub fn destroy(self: Self) void {
+    if(builtin.is_test) {
+        return;
+    }
     self.window.destroy();
 }
 
 /// Call `Window.terminate()` to clear resources
 pub fn terminate() void {
+    if(builtin.is_test) {
+        return;
+    }
     glfw.terminate();
 }
 
@@ -138,9 +162,24 @@ const Err = error{
 };
 
 const std = @import("std");
-const glfw = @import("glfw");
-const gl = @import("gl");
+const builtin = @import("builtin");
+const glfw = blk: {
+    if (builtin.is_test) {
+        break :blk EmptyStruct;
+    } else {
+        break :blk @import("glfw");
+    }
+};
+const gl = blk: {
+    if (builtin.is_test) {
+        break :blk EmptyStruct;
+    } else {
+        break :blk @import("gl");
+    }
+};
 const gfx = @import("graphics.zig");
 const ft = @import("ft.zig");
 const Context = @import("Context.zig");
 const Component = @import("ui/ui.zig").Component;
+const App = @import("App.zig");
+const EmptyStruct = @import("testing.zig").EmptyStruct;
